chore: remove subscription
