+++ pages/account.tsx
@@ -116,7 +116,7 @@ export default function Account({ user }: { user: User }) {
             </div>
           }
         >
-          <div className="text-xl mt-8 mb-4 font-semibold">
+          {/* <div className="text-xl mt-8 mb-4 font-semibold">
             {isLoading ? (
               <div className="h-12 mb-6">
                 <LoadingDots />
@@ -126,7 +126,7 @@ export default function Account({ user }: { user: User }) {
             ) : (
               <Link href="/">Choose your plan</Link>
             )}
-          </div>
+          </div> */}
         </Card>
         <Card
           title="Your Name"
@@ -134,7 +134,8 @@ export default function Account({ user }: { user: User }) {
           footer={<p>Please use 64 characters at maximum.</p>}
         >
           <div className="text-xl mt-8 mb-4 font-semibold">
-            {userDetails ? (
+            {user.user_metadata.full_name}
+            {/* {userDetails ? (
               `${
                 userDetails.full_name ??
                 `${userDetails.first_name} ${userDetails.last_name}`
@@ -143,7 +144,7 @@ export default function Account({ user }: { user: User }) {
               <div className="h-8 mb-6">
                 <LoadingDots />
               </div>
-            )}
+            )} */}
           </div>
         </Card>
         <Card
