+++ pages/index.tsx
@@ -1,23 +1,23 @@
 import { GetStaticPropsResult } from 'next';
 
-import Pricing from '@/components/Pricing';
+import Home from '@/components/Home';
 import { getActiveProductsWithPrices } from '@/utils/supabase-client';
 import { Product } from 'types';
 
-interface Props {
-  products: Product[];
+// interface Props {
+//   products: Product[];
+// }
+
+export default function PricingPage() {
+  return <Home />;
 }
 
-export default function PricingPage({ products }: Props) {
-  return <Pricing products={products} />;
-}
-
-export async function getStaticProps(): Promise<GetStaticPropsResult<Props>> {
-  const products = await getActiveProductsWithPrices();
+export async function getStaticProps() {
+  //const products = await getActiveProductsWithPrices();
 
   return {
     props: {
-      products
+      products: null,
     },
     revalidate: 60
   };
